(Refactor): Kafka consume 수동 커밋 진행으로 인한 ack 인자 추가
