+++ src/main/resources/application.yml
@@ -23,6 +23,7 @@ spring:
     listener:
       missing-topics-fatal: false # 존재하는 토픽으로 인한 앱 실패 방지 설정
 
+
   redis:
     host: localhost
     port: 6379
